Made SortedVec an abstract class and introduced an OrderedNonces case class
