+++ core/src/main/scala/org/bitcoins/core/util/sorted/OrderedNonces.scala
@@ -1,0 +1,8 @@
+package org.bitcoins.core.util.sorted
+
+import org.bitcoins.crypto.SchnorrNonce
+
+/** Represents an ordered set of SchnorrNonces */
+case class OrderedNonces(vec: Vector[SchnorrNonce])
+    extends SortedVec[SchnorrNonce, SchnorrNonce](vec,
+                                                  SortedVec.forOrdered(vec))
